web: components: Moved story for Text + added base test
