--- web/ASC.Web.Storybook/stories/text/content-header/index.stories.js
@@ -1,26 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-import { text, boolean, withKnobs, color } from '@storybook/addon-knobs/react';
-import { Text } from 'asc-web-components';
-import Section from '../../../.storybook/decorators/section';
-import withReadme from 'storybook-readme/with-readme';
-import Readme from './README.md';
-
-storiesOf('Components|Text', module)
-  .addDecorator(withKnobs)
-  .addDecorator(withReadme(Readme))
-  .add('content header', () => (
-
-    <Section>
-      <div style={{ width: "100%" }}>
-        <Text.ContentHeader
-          color={color('color', '#333333')}
-          title={text('title', '')}
-          truncate={boolean('truncate', false)}
-          isInline={boolean('isInline', false)}
-        >
-          {text('Text', 'Sample text Headline')}
-        </Text.ContentHeader>
-      </div>
-    </Section>
-  ));
