--- web/ASC.Web.Storybook/stories/text/menu-header/README.md
@@ -1,28 +1,0 @@
-# Text
-
-## Usage
-
-```js
-import { Text } from 'asc-web-components';
-```
-
-### <Text.MenuHeader>
-
-Wraps the given text in the specified size of the menu header.
-
-#### Usage
-
-```js
-    <Text.MenuHeader title='Some title' isInline>
-        Some text
-    </Text.MenuHeader>
-```
-
-#### Properties
-
-| Props              | Type     | Required | Values                      | Default   | Description                                                                                                                                      |
-| ------------------ | -------- | :------: | --------------------------- | --------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `color`            | `string` |    -     | -                     | '#333333' | Specifies the menuHeader color                      |
-| `title`            | `bool`   |    -     | -                     | -         | Title                                               |
-| `truncate`         | `bool`   |    -     | -                     | false     | Disables word wrapping                              |
-| `isInline`         | `bool`   |    -     | -                     | false     | Sets the 'display: inline-block' property           |
